Update for refactor on kinvey-node-sdk
