+++ .babelrc
@@ -1,6 +1,12 @@
 {
   "plugins": [
-    "inline-dotenv"
+    "inline-dotenv",
+    ["module-resolver", {
+      "alias": {
+        "package.json": "./package.json",
+        "src": "./src"
+      }
+    }]
   ],
   "presets": [
     "es2015",
