+++ .eslintrc
@@ -37,5 +37,10 @@
     "es6": true,
     "mocha": true,
     "node": true
+  },
+  "settings": {
+    "import/resolver": {
+      "babel-module": {}
+    }
   }
 }
