+++ package.json
@@ -39,18 +39,19 @@
     "version": "npm run build && git add -A dist"
   },
   "dependencies": {
-    "core-js": "^2.4.1",
-    "es6-promise": "^4.0.5",
-    "kinvey-html5-sdk": "3.3.5",
-    "kinvey-node-sdk": "3.3.5",
-    "local-storage": "^1.4.2",
-    "lodash": "^4.8.2"
+    "core-js": "2.4.1",
+    "es6-promise": "4.0.5",
+    "kinvey-html5-sdk": "file:///Users/Thomas/Documents/Kinvey/Development/SDKs/JavaScript/HTML5/SDK",
+    "kinvey-node-sdk": "file:///Users/Thomas/Documents/Kinvey/Development/SDKs/JavaScript/Node/SDK",
+    "local-storage": "1.4.2",
+    "lodash": "4.17.4"
   },
   "devDependencies": {
     "babel-cli": "^6.16.0",
     "babel-core": "^6.9.0",
     "babel-eslint": "^6.0.0",
     "babel-plugin-inline-dotenv": "^1.1.1",
+    "babel-plugin-module-resolver": "2.4.0",
     "babel-polyfill": "^6.9.0",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-stage-2": "^6.0.15",
@@ -59,6 +60,7 @@
     "del-cli": "^0.2.0",
     "eslint": "^3.5.0",
     "eslint-config-airbnb-base": "^7.1.0",
+    "eslint-import-resolver-babel-module": "2.2.1",
     "eslint-plugin-import": "^1.15.0",
     "expect": "1.20.2",
     "fs-extra": "^0.30.0",
