--- src/export.js
@@ -1,1 +1,0 @@
-export * from 'kinvey-html5-sdk/dist/export';
