+++ src/middleware/index.js
@@ -1,0 +1,5 @@
+import HttpMiddleware from './src/http';
+
+export {
+  HttpMiddleware
+};
