+++ webpack.config.js
@@ -16,7 +16,7 @@ var BANNER = '/**\n'
 
 module.exports = {
   context: path.resolve(__dirname, 'dist'),
-  entry: ['core-js/es6/symbol', './index.js'],
+  entry: ['core-js/es6/symbol', 'es6-promise/auto', './index.js'],
   module: {
     loaders: [
       { test: /\.json$/, loader: 'json-loader' }
@@ -26,21 +26,13 @@ module.exports = {
     filename: pkg.name + '.js',
     libraryTarget: 'umd',
     library: 'Kinvey',
-    path: path.resolve(__dirname, 'dist')
+    path: path.resolve(__dirname, './dist')
   },
   plugins: [
     new webpack.BannerPlugin(BANNER, { raw: true }),
     new webpack.NormalModuleReplacementPlugin(
-      /kinvey-node-sdk\/dist\/request\/src\/device\.js/,
-      require.resolve(path.resolve(__dirname, 'dist/device.js'))
-    ),
-    new webpack.NormalModuleReplacementPlugin(
       /kinvey-node-sdk\/dist\/identity\/src\/popup\.js/,
-      require.resolve(path.resolve(__dirname, 'dist/popup.js'))
-    ),
-    new webpack.NormalModuleReplacementPlugin(
-      /kinvey-node-sdk\/dist\/request\/src\/middleware\/src\/cache\.js/,
-      require.resolve(path.resolve(__dirname, 'node_modules/kinvey-html5-sdk/dist/middleware/src/cache.js'))
+      require.resolve(path.resolve(__dirname, './dist/popup.js'))
     )
   ]
 };
